move ClosingStateChange to its own file
